✨ Fixing sonarqube bugs
